remove apiData reference from searchCatalogSchema
also moved catalogSchema to a common location since multiple server
schemas use it
