+++ src/lib/server/schema/queryAvailableProgramsSchema.ts
@@ -1,14 +1,5 @@
 import { z } from 'zod';
-
-// TODO: move to a more central place?
-export const catalogSchema = z
-  .string({
-    required_error: 'Catalog is required.'
-  })
-  .refine((catalog) => {
-    const parts = catalog.split('-').map((part) => Number(part));
-    return catalog.length === 9 && !isNaN(parts[0]) && !isNaN(parts[1]) && parts[1] > parts[0];
-  }, 'Invalid catalog format.');
+import { catalogSchema } from '$lib/server/schema/common';
 
 // validation schema for api/data/queryAvailablePrograms endpoint
 // currently only support the following query types:
